BookResourceImpl is updated and added some conditions which handles exception

+++ src/main/java/com/webservices/restWebServices/resources/BookResourceImpl.java
@@ -5,7 +5,6 @@ import com.webservices.restWebServices.model.Book;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.sql.DataSource;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
 import java.util.List;
 
@@ -18,6 +17,7 @@ public class BookResourceImpl implements BookResource{
     @Autowired
     private BookRepository bookRepository;
 
+    //method to add book
     @Override
     public Response addBook(Book book) {
         Book bookExists = bookRepository.findByIsbn(book.getIsbn());
@@ -27,37 +27,48 @@ public class BookResourceImpl implements BookResource{
             bookRepository.save(book);
             return Response.ok("Book with ISBN " + book.getIsbn() + " added successfully.").build();
         }
-
     }
 
+    //method to show all the books added
     @Override
     public List<Book> getAllBooks() {
         return bookRepository.findAll();
     }
 
+    //method to show book of particular isbn
     @Override
-    public Book getBookByIsbn(@PathParam("isbn") Long isbn) {
+    public Book getBookByIsbn(Long isbn) {
         return bookRepository.findByIsbn(isbn);
     }
 
+    //method to delete book
+    //isbn is used to decide which book to delete
     @Override
-    public Response deleteBookByIsbn(@PathParam("isbn") Long isbn) {
+    public Response deleteBookByIsbn(Long isbn) {
         Book isbnExistence = bookRepository.findByIsbn(isbn);
         System.out.println(isbnExistence);
-//        if(isbnExistence == null){
-//            return Response.status(Response.Status.NOT_FOUND).entity("ISBN is not found").build();
-//        } else {
+        if(isbnExistence == null){
+            return Response.status(Response.Status.NOT_FOUND).entity("ISBN is not found").build();
+        } else {
             bookRepository.deleteById(isbn);
-            return Response.ok("Book with ISBN " + isbn + " Deleted Successfully.").build();
-//        }
+            return Response.ok("Book with ISBN " + isbn + " deleted Successfully.").build();
+        }
 
     }
 
+    //update book details
     @Override
-    public Response updateVintage(@PathParam("isbn") Long isbn, Book book) {
-        book.setIsbn(isbn);
-        bookRepository.save(book);
-        return Response.ok("Book with ISBN " +isbn+ " is updated. " ).build();
+    public Response updateBook(Long isbn, Book book) {
+        Book isbnExistence = bookRepository.findByIsbn(isbn);
+        System.out.println(isbnExistence);
+        if(isbnExistence == null){
+            return Response.status(Response.Status.NOT_FOUND).entity("ISBN is not found").build();
+        } else {
+            book.setIsbn(isbn);
+            bookRepository.save(book);
+            return Response.ok("Book with ISBN " +isbn+ " is updated. " ).build();
+        }
+
     }
